Ignore implicit ctors in anonymous structs.
For anonymous structs Clang generates no name for its implicit ctor. This
was confusing the profile reader in LLVM. Since we can't really map
an empty name to a callgraph node, this patch simply ignores anonymous
structs.

This should not be an issue. If an anonymous struct has a substantial
amount of cycles in its ctor, so will its parent.

Additonally, this patch makes the profile converter ignore functions
for which we have no locations with non-zero samples. They are not
worth writing out.