Generate a binary-encoded profile for LLVM profiles.
This changes the LLVM profile writer to use a more compact binary
encoding, as supported in upstream LLVM as of

http://llvm.org/viewvc/llvm-project?rev=250539&view=rev

This change breaks compatibility with previous versions. To dump the
generated file, you will need to use the llvm-profdata tool in LLVM.
Make sure to get a version of llvm-profdata at at least revision 250539.