Add configuration support to detect and use LLVM libraries.
This adds the --with-llvm configuration flag.  It's used to find
the path to LLVM's llvm-config which produces the pre-processor and
linker flags needed to include LLVM's ProfileData library.

This change only includes the necessary configury bits to support
building with LLVM libraries.  It does not change the LLVM profile
converter at all.