Support for cache prefetching profiles. (#75)
Recommendations for locations to insert software cache prefetches are provided in a csv file as a triple `binary offset`,`delta`,`type`.

This is part of a larger system, consisting of a cache prefetches recommender, create_llvm_prof, and clang.

An example recommender is DynamoRIO's cache miss analyzer. It processes memory access traces obtained from a running binary and identifies patterns in cache misses. Based on them, it produces a csv file with recommendations. A recommendation is made in terms of: 

- the value of a memory operand of an instruction at a certain binary offset; 
- a delta; 
- and a type (nta, t0, t1, t2)

meaning: a prefetch of that type should be inserted right before the instrution at that binary offset, and the prefetch should be for an address delta away from the memory address the instruction will access. 

For example:

`0x400ab2,64,nta`

and assuming the instruction at 0x400ab2 is:

`movzbl (%rbx,%rdx,1),%edx`

means that the recommender determined it would be beneficial for a prefetchnta instruction to be inserted right before this instruction, as such:

```
prefetchnta 0x40(%rbx,%rdx,1)
movzbl (%rbx, %rdx, 1), %edx
```

See https://goo.gl/6TM2Xp for a script demonstrating this.

The workflow for prefetch cache instrumentation is as follows:

1. build binary.
2. collect memory traces, run analysis to obtain recommendations
3. convert recommendations to afdo using create_llvm_prof:

`create_llvm_prof --binary=<binary> --profile=<recommendations.csv> --profiler="prefetch" --out=<some_file.afdo> --format=<text or binary>`

4. rebuild binary, using the exact same set of arguments used initially, to which `-mllvm -prefetch-hints-file=<file>` needs to be added (the later is part of the upcoming LLVM change)

Note that if sample profiling feedback-driven optimization is desired, that happens before step 1 above. In this case, the sample profile afdo file that was used to produce the binary at step 1 must also be included in step 4.

The role of this change is to convert prefetching recommendations into data the compiler can use. This requires mapping the binary offsets to debug information. An upcoming LLVM change ensures each memory operand instruction is uniquely identified through debug info.

This is very similar to what needs to be done for sample profiles. For this reason, and given that the overall approach (memory tracing-based cache recommendation mechanisms) is under active development, we use the afdo format as a syntax for capturing this information. We avoid confusing semantics with sample profile afdo data by feeding the two types of information to the compiler through separate files and compiler flags.