Fix LLVM profile merger flags
Currently the LLVM profile merger interprets all the flags passed on
the command line as input files. This makes it impossible to pass flags
into the executable as they will be interpreted as files and then the
profile reader will fail due to the file not existing. This patch fixes
that behavior by taking the leftover positional arguments from the absl
flag parsing function.