Avoid breakages for builds with libprotobuf < 3.16.0 (#197)
Recent commit 623c777 upgraded perf_data_converter to a version that
uses `google::protobuf::RepeatedField<unsigned int>::Assign`, which was
introduced in libprotobuf version 3.16.0. This commit breaks builds
that use libprotobuf < 3.16.0, such as stock builds on Ubuntu 20.04
(which uses v3.6.1) and Ubuntu 22.04 (which uses v3.12.4).

Thankfully, this dependency was introduced after the SPE Decoder, so we
can take a snapshot of perf_data_converter in the compatible range.