Write out file names to GCOV file
This patch modifies the profile readers and writers to support writing
out the file names for functions into the GCOV file to enable using
file names for more accurate profile annotation in GCC's auto-profile
pass. This patch goes along with the RFC posted to the GCC mailing lists at
https://gcc.gnu.org/pipermail/gcc-patches/2025-June/686835.html.

This also adds support for GCOV version 3, to support the breaking
change in the file format.